Move device entry tests to test_init and avoid snapshot list
Suggested-by: joostlek
